# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * YouTube Automation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  niche?: string | null;
  /** @nullable */
  language?: string | null;
  status: ProjectStatus;
  videoCount?: number;
  uploadedCount?: number;
  createdAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  niche?: string;
  language?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  paused: 'paused',
  archived: 'archived',
} as const;

export interface ProjectUpdate {
  name?: string;
  description?: string;
  niche?: string;
  language?: string;
  status?: ProjectUpdateStatus;
}

export type VideoStatus = typeof VideoStatus[keyof typeof VideoStatus];


export const VideoStatus = {
  pending: 'pending',
  generating_script: 'generating_script',
  generating_voice: 'generating_voice',
  generating_thumbnail: 'generating_thumbnail',
  assembling: 'assembling',
  ready: 'ready',
  uploading: 'uploading',
  uploaded: 'uploaded',
  failed: 'failed',
} as const;

export interface Video {
  id: number;
  projectId: number;
  topic: string;
  /** @nullable */
  script?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  seoTags?: string | null;
  status: VideoStatus;
  /** @nullable */
  pipelineStep?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  youtubeId?: string | null;
  /** @nullable */
  youtubeUrl?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  audioPath?: string | null;
  /** @nullable */
  videoPath?: string | null;
  /** @nullable */
  durationSeconds?: number | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface VideoInput {
  projectId: number;
  /** @minLength 1 */
  topic: string;
  language?: string;
}

export interface AiStatus {
  ollamaAvailable: boolean;
  ollamaUrl?: string;
  models?: string[];
  ttsAvailable: boolean;
  ttsEngine?: string;
}

export interface Settings {
  id: number;
  ollamaUrl?: string;
  ollamaModel?: string;
  ttsEngine?: string;
  youtubeApiKeySet?: boolean;
  /** @nullable */
  youtubeChannelId?: string | null;
  defaultLanguage?: string;
  autoUpload?: boolean;
  videosOutputDir?: string;
}

export interface SettingsUpdate {
  ollamaUrl?: string;
  ollamaModel?: string;
  ttsEngine?: string;
  youtubeApiKey?: string;
  youtubeChannelId?: string;
  defaultLanguage?: string;
  autoUpload?: boolean;
  videosOutputDir?: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalVideos: number;
  uploadedVideos: number;
  pendingVideos: number;
  failedVideos: number;
  recentVideos?: Video[];
}

export type ListVideosParams = {
projectId?: number;
status?: string;
};