# lib/api-zod/src/generated/types/aiStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * YouTube Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface AiStatus {
  ollamaAvailable: boolean;
  ollamaUrl?: string;
  models?: string[];
  ttsAvailable: boolean;
  ttsEngine?: string;
}