# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * YouTube Automation System API
 * OpenAPI spec version: 0.1.0
 */
import type { Video } from './video';

export interface DashboardStats {
  totalProjects: number;
  totalVideos: number;
  uploadedVideos: number;
  pendingVideos: number;
  failedVideos: number;
  recentVideos?: Video[];
}