# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * YouTube Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  niche?: string;
  language?: string;
}